This represents my changes for week #8
Code changes that focused primarily on stories CT-103, CT-111, and CT-135 around tying the UI to retrieve DB data and return textual and graphical results. Much of the code in the views.py is still quasi POC with regard to dynamic image creation and rendering. I had originally planned on using DB views and dynamically filter them with a "where" clause, but presented significant performance issues. I am now trying to use dynamic SQL, which is not working yet in the code (apparent python/mysql syntax error) but should be fixed by Monday 10/19/2020.