Incremental updates as the class framework development continues
I have finished and unit tested tge new models classes that are part of the framework. The views.py is being checked into as a reference for testing I have done as I started to retire the original code from views.py. I tested all of the UI functions and they seem to work correctly for the cases of zipcode has no data, a single county, or multiple counties.