Graphics class framework, supporting updates in models and views
The graphics class framework has been created and supporting code in models and views is being checked in, too. The views.py code is being used as a test bed for the code changes. I will be working on the controller pieces tomorrow and hope to have them checked in sometime in the afternoon.